src/requirements.txt
appnope==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
en-core-web-sm==1.2.0
entrypoints==0.2.3
ftfy==4.4.3
gensim==2.3.0
h5py==2.7.0
html5lib==0.9999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.6
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
murmurhash==0.26.4
nbconvert==5.2.1
nbformat==4.3.0
nltk==3.2.4
notebook==5.0.0
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pandocfilters==1.4.1
pathlib==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
plac==0.9.6
preshed==1.0.0
progressbar2==3.33.2
prompt-toolkit==1.0.15
protobuf==3.3.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
regex==2017.7.28
requests==2.18.2
scikit-learn==0.18.2
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
six==1.10.0
smart-open==1.5.3
spacy==1.9.0
tensorflow==1.2.1
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
thinc==6.5.2
toolz==0.8.2
torch==0.2.0.post1
torchvision==0.1.9
tornado==4.5.1
tqdm==4.15.0
traitlets==4.3.2
ujson==1.35
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wrapt==1.10.10
xlrd==1.0.0
